backend_flask/requirements.txt
Flask
flask-cors
python-dotenv
PyJWT
bcrypt
PyMySQL
